Simplify CommittedAtBackfiller.
